Lots of Maths - Tweaked Seeking Projectile Class
Created A Prototype Actor to visualise some of the math calculations.
Can now detect if a point is on the right / left side of you respectively.
Can also translate a direction vector into its vertical and horizontal components + magnitudes.
Can now map a ratio of values to a different ratio of values.